feat(console): migrate "Documentation" into new menu for API v2
