+++ gravitee-apim-console-webui/src/components/documentation/documentation-management.html
@@ -20,7 +20,7 @@
     <div layout="row" layout-align="space-between center" ng-if="!renameFolder">
       <h3>
         <span ng-repeat="folder in $ctrl.breadcrumb">
-          <a ng-if="$ctrl.apiId && !$last" ui-sref="management.apis.detail.portal.documentation({apiId: $ctrl.apiId, parent: folder.id})"
+          <a ng-if="$ctrl.apiId && !$last" ui-sref="management.apis.ng.documentation({apiId: $ctrl.apiId, parent: folder.id})"
             >{{folder.name}}</a
           >
           <a ng-if="!$ctrl.apiId && !$last" ui-sref="management.settings.documentation.list({parent: folder.id})">{{folder.name}}</a>
