+++ gravitee-apim-console-webui/src/components/documentation/new-page.component.ts
@@ -14,185 +14,39 @@
  * limitations under the License.
  */
 
-import { StateService } from '@uirouter/core';
-import { IController, IPromise, IScope } from 'angular';
-import * as _ from 'lodash';
+import { Component, ElementRef, Injector, Input, SimpleChange } from '@angular/core';
+import { UpgradeComponent } from '@angular/upgrade/static';
 
-import { DocumentationQuery, DocumentationService, PageType } from '../../services/documentation.service';
-import NotificationService from '../../services/notification.service';
-
-interface IPageScope extends IScope {
-  getContentMode: string;
-  fetcherJsonSchema: string;
-}
-class NewPageComponentController implements IController {
-  resolvedFetchers: any[];
-  folders: any[];
-  systemFolders: any[];
-  pageResources: any[];
-  categoryResources: any[];
-  pagesToLink: any[];
-
-  apiId: string;
-  error: any;
-  page: any;
-  foldersById: _.Dictionary<any>;
-  systemFoldersById: _.Dictionary<any>;
-  pageList: any[];
-  templates: any[];
-  selectedTemplate: any;
-  /* @ngInject */
-  constructor(
-    private readonly NotificationService: NotificationService,
-    private readonly DocumentationService: DocumentationService,
-    private readonly Constants,
-    private $state: StateService,
-    private $scope: IPageScope,
-  ) {
-    this.apiId = this.$state.params.apiId;
-    this.error = null;
-    this.page = {
-      name: '',
-      type: this.$state.params.type,
-      parentId: this.$state.params.parent,
-      visibility: 'PUBLIC',
-    };
-
-    this.$scope.getContentMode = 'inline';
+@Component({
+  template: '',
+  selector: 'documentation-new-page',
+  host: {
+    class: 'bootstrap',
+  },
+})
+export class DocumentationNewPageComponent extends UpgradeComponent {
+  @Input() resolvedFetchers;
+  @Input() folders;
+  @Input() systemFolders;
+  @Input() pageResources;
+  @Input() categoryResources;
+  @Input() pagesToLink;
+  constructor(elementRef: ElementRef, injector: Injector) {
+    super('documentationNewPageAjs', elementRef, injector);
   }
 
-  $onInit() {
-    this.foldersById = _.keyBy(this.folders, 'id');
-    this.systemFoldersById = _.keyBy(this.systemFolders, 'id');
-    const folderSituation = this.DocumentationService.getFolderSituation(this.systemFoldersById, this.foldersById, this.page.parentId);
-    this.pageList = this.DocumentationService.buildPageList(this.pageResources, true, folderSituation);
-    this.pagesToLink = this.DocumentationService.buildPageList(this.pagesToLink, false, folderSituation);
-
-    if (this.DocumentationService.supportedTypes(folderSituation).indexOf(this.page.type) < 0) {
-      this.$state.go('management.settings.documentation.list', { parent: this.$state.params.parent });
-    }
-
-    const q = new DocumentationQuery();
-    q.type = PageType.MARKDOWN_TEMPLATE;
-    q.published = true;
-    q.translated = true;
-    this.DocumentationService.search(q, null).then((response) => {
-      this.templates = response.data;
+  ngOnInit() {
+    // Hack to Force the binding between Angular and AngularJS
+    // Don't know why, but the binding is not done automatically when resolver is used
+    this.ngOnChanges({
+      resolvedFetchers: new SimpleChange(null, this.resolvedFetchers, true),
+      folders: new SimpleChange(null, this.folders, true),
+      systemFolders: new SimpleChange(null, this.systemFolders, true),
+      pageResources: new SimpleChange(null, this.pageResources, true),
+      categoryResources: new SimpleChange(null, this.categoryResources, true),
+      pagesToLink: new SimpleChange(null, this.pagesToLink, true),
     });
 
-    const settings = this.Constants.env.settings;
-    if (this.page.type === 'SWAGGER' && settings && settings.openAPIDocViewer) {
-      this.page.configuration = {
-        viewer: settings.openAPIDocViewer.openAPIDocType.defaultType,
-      };
-    }
-  }
-
-  getPageName(): string {
-    switch (this.page.type) {
-      case PageType.ASCIIDOC:
-        return 'New AsciiDoc';
-      case PageType.ASYNCAPI:
-        return 'New AsyncApi';
-      case PageType.FOLDER:
-        return 'New Folder';
-      case PageType.LINK:
-        return 'New Link';
-      case PageType.MARKDOWN_TEMPLATE:
-        return 'New Markdown Template';
-      case PageType.MARKDOWN:
-        return 'New Markdown Page';
-      case PageType.SWAGGER:
-        return 'New Swagger Template';
-      default:
-        return 'New Page';
-    }
-  }
-
-  isFolder(): boolean {
-    return PageType.FOLDER === this.page.type;
-  }
-  isLink(): boolean {
-    return PageType.LINK === this.page.type;
-  }
-  isSwagger(): boolean {
-    return PageType.SWAGGER === this.page.type;
-  }
-  isMarkdown(): boolean {
-    return PageType.MARKDOWN === this.page.type;
-  }
-  isMarkdownTemplate(): boolean {
-    return PageType.MARKDOWN_TEMPLATE === this.page.type;
-  }
-
-  onChangeMarkdownTemplate() {
-    if (this.selectedTemplate.type) {
-      this.page = { ...this.page, content: this.selectedTemplate.content };
-    }
-  }
-
-  save(gotoParent: boolean): IPromise<void> {
-    this.error = null;
-    return this.DocumentationService.create(this.page, this.apiId)
-      .then((response: any) => {
-        const page = response.data;
-        if (page.messages && page.messages.length > 0) {
-          this.NotificationService.showError("'" + page.name + "' has been created (with validation errors)");
-        } else {
-          this.NotificationService.show("'" + page.name + "' has been created");
-        }
-        if (gotoParent) {
-          this.gotoParent();
-        } else {
-          this.gotoEdit(page);
-        }
-      })
-      .catch((err) => {
-        this.error = { ...err.data, title: 'Sorry, unable to create page' };
-      });
-  }
-
-  changeContentMode(newMode) {
-    if ('fetcher' === newMode) {
-      this.page.source = {
-        configuration: {},
-      };
-    } else {
-      delete this.page.source;
-    }
-    this.error = null;
-  }
-
-  cancel() {
-    this.gotoParent();
-  }
-
-  gotoParent() {
-    if (this.apiId) {
-      this.$state.go('management.apis.detail.portal.documentation', { apiId: this.apiId, parent: this.$state.params.parent });
-    } else {
-      this.$state.go('management.settings.documentation.list', { parent: this.$state.params.parent });
-    }
-  }
-
-  gotoEdit(page: any) {
-    if (this.apiId) {
-      this.$state.go('management.apis.detail.portal.editdocumentation', { apiId: this.apiId, pageId: page.id, type: page.type });
-    } else {
-      this.$state.go('management.settings.documentation.edit', { pageId: page.id, type: page.type });
-    }
+    super.ngOnInit();
   }
 }
-
-export const NewPageComponent: ng.IComponentOptions = {
-  bindings: {
-    resolvedFetchers: '<',
-    folders: '<',
-    systemFolders: '<',
-    pageResources: '<',
-    categoryResources: '<',
-    pagesToLink: '<',
-  },
-  template: require('./new-page.html'),
-  controller: NewPageComponentController,
-};
